Use specific classes for pointer args. Unused at the moment.
This could be used to create  simplier to use pointer logic. Since the release event only returns the release location, it is on the programmer to keep track of the initial click location. Same applies to the move event. In a game like brick breaker, the programmer will handle all events as they need to display the current aimed location as the user pans around. Other games might not need to track the pointer down event and would be fine with a single pointer released event that includes the initial down location. Still need to figure out how to get that info into the PointerReleasedEvent efficently.